added user profile app and model
