+++ backend/clubscheduler-backend/profile_api/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import UserProfile
+
+# Register your models here.
+
+admin.site.register(UserProfile)
