+++ backend/clubscheduler-backend/profile_api/views.py
@@ -1,6 +1,3 @@
 from django.shortcuts import render
 
 # Create your views here.
-
-
-
