+++ backend/clubscheduler-backend/user_api/apps.py
@@ -4,5 +4,5 @@ class UserApiConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "user_api"
 
-    def ready(self):
-        import user_api.signals
+    # def ready(self):
+    #     import user_api.signals
